rendering: add a new function to customize DataFrame
* Implement the function style_dataframe() in utils_streamlit_app.py
to customize the rendering of DataFrames.

* The function uses CSS selectors to target specific elements, such as
rows, columns, and cells, and modify their CSS properties.

* Note that this approach may change in the future when a versions of
Streamlit provide more customization options for DataFrames.
